Make table of rates given a denominator

calc_shares makes it easy to divide values by some denominator within the same long-shaped data frame. For example, it works well for a table of population groups for multiple locations where you want to divide population counts by some total population. It optionally handles dividing margins of error. Denote locations or other groupings by using a grouped data frame, passing bare column names to ..., or both.

calc_shares(df, ..., group = group, denom = "Total", estimate = estimate,
  moe = NULL)

Arguments

df
A data frame

...
Optional; bare column names to be used for groupings.

group
Bare column name where groups are given--that is, the denominator value should be found in this column

denom
String; denominator to filter from group

estimate
Bare column name of estimates or values

moe
Bare column name of margins of error; if supplied, MOE of shares will be included in output

Value

A tibble/data frame with shares (and optionally MOE of shares) of subgroup values within a denominator group. Shares given for denominator group will be blank.

Examples

edu %>%
  dplyr::group_by(name) %>%
  calc_shares(group = variable, denom = "age25plus", moe = moe)

#> # A tibble: 15 x 6
#> # Groups:   name [5]
#>    name       variable  estimate   moe  share sharemoe
#>    <chr>      <fct>        <dbl> <dbl>  <dbl>    <dbl>
#>  1 Bethany    age25plus     4013   148 NA       NA    
#>  2 Bethany    bachelors      994   174  0.248    0.042
#>  3 Bethany    masters        655   153  0.163    0.038
#>  4 East Haven age25plus    21230   490 NA       NA    
#>  5 East Haven bachelors     2998   361  0.141    0.017
#>  6 East Haven masters       1280   229  0.06     0.011
#>  7 Hamden     age25plus    40110   721 NA       NA    
#>  8 Hamden     bachelors     8587   669  0.214    0.016
#>  9 Hamden     masters       6171   568  0.154    0.014
#> 10 New Haven  age25plus    81231  1252 NA       NA    
#> 11 New Haven  bachelors    12582   820  0.155    0.01 
#> 12 New Haven  masters       8236   582  0.101    0.007
#> 13 West Haven age25plus    36031   765 NA       NA    
#> 14 West Haven bachelors     4865   533  0.135    0.015
#> 15 West Haven masters       2267   388  0.063    0.011

race_pops %>%
  calc_shares(region, name, group = variable, denom = "total", moe = moe)

#> # A tibble: 65 x 7
#>    region     name       variable estimate   moe  share sharemoe
#>    <fct>      <chr>      <fct>       <dbl> <dbl>  <dbl>    <dbl>
#>  1 New Haven  New Haven  total      130405    60 NA       NA    
#>  2 New Haven  New Haven  white       40164  1395  0.308    0.011
#>  3 New Haven  New Haven  black       42970  1383  0.33     0.011
#>  4 New Haven  New Haven  asian        6042   697  0.046    0.005
#>  5 New Haven  New Haven  latino      37231  1688  0.286    0.013
#>  6 Inner Ring East Haven total       29015    24 NA       NA    
#>  7 Inner Ring East Haven white       22898   785  0.789    0.027
#>  8 Inner Ring East Haven black         693   400  0.024    0.014
#>  9 Inner Ring East Haven asian        1123   595  0.039    0.021
#> 10 Inner Ring East Haven latino       3820   641  0.132    0.022
#> # ... with 55 more rows