Theme Din

A clean theme especially suited for labeled bar charts or Cleveland dot plots. Designed based on the DIN family of fonts, but now defaults to font "Roboto Condensed". This defaults to base_family = "roboto", which depends on a) using showtext and sysfonts to get a font called "roboto", or b) already having a font loaded with this name. Use sysfonts::font_add_google or sysfonts::font_add to set a different font.

theme_din(base_size = 14, base_family = "roboto", xgrid = FALSE,
  ygrid = TRUE, fallback_google = TRUE)

Arguments

base_size
Base font size

base_family
Base font family; defaults to "roboto", as set by showtext

xgrid
A logical for turning x-grid on or off, or "dotted", for a light dotted grid

ygrid
A logical for turning y-grid on or off, or "dotted", for a light dotted grid

fallback_google
Logical: if TRUE and base_family not currently loaded, will load Roboto Condensed from Google. If FALSE, will load system sans font. Defaults TRUE.

See also